+++ ElvUI_MerathilisUI/Core/Changelog/7.31.lua
@@ -10,6 +10,7 @@ MER.Changelog[731] = {
 	},
 	NEW = {
 		"[Options]: Module \"Enable\" toggle switches now color their label green/red to reflect the on/off state",
+		"[Skins]: Added font size/outline options for the Weekly Rewards (Great Vault) skin (#124)",
 	},
 	IMPROVEMENTS = {},
 }